Gather imports from used types
Instead of always importing all other modules, and relying on the
autoformatter to remove the unused ones, gather the modules that need to
be imported by examining the types that are being used.

Ruff doesn't seem to be very successful to remove the unused imports. It
does to an extent, for example removing the unused imported types from
the `typing` module, or the `os` or `enum` module if they aren't used;
but it failed to remove imports of unused first-party modules that
aren't from the standard library.

We need the imports to resolve types, so use `python_type` to examine
what modules are required and add them to a list.