Add support for type aliases and preambles
This changes allows PyExportTraits to define `py_typing_preamble` that
contains some Python code that must be included at the top of the
module. It can be a multiline string of Python code, but it can only
contain staments starting with:

  - `import`: to import whole modules
  - `from`: to import specific types from a module
  - `type`: to declare a type alias [1]

The set of `imports` gathered from `python_type` has been generalized to
a list of `preamble` statements that can contain both imports and type
aliases.

For example, this helps us to type-hint `prim::Point3D<float>` as
`_Point3D[float]` instead of `tuple[float, float, float]`; or to
type-hint `std::filesystem::path` as `StrOrBytesPath` imported from
`_typeshed`.

[1] https://typing.python.org/en/latest/spec/aliases.html#type-statement