Support different type hints for parameters and return values
Add support for `py_typing_param` as alternative type-hint for
parameters, and add `T!` syntax to prevent it from being used for
template substitutions.

As a parameter, there may be some implicit conversions accepted from
other types that will never be used as the return type. Using the same
type-hint for both parameters and return values has some limits: either
the parameter type is to strict (and not accepted the conversions), or
the return type is too broad.

For example, `std::filesystem::path` accepts all of `StrOrBytesPath`
(`str`, `bytes` or anything with `__fspath__` convertor, including
`pathlib.Path`), but it only ever returns `pathlib.Path` instances.
Before this change, we annotate the return type too broadly.

`float` as return value or constant is always `float`, but as parameter
it accepts anything that has `__float__` or `__index__` convertors.
Before this change, we annotate the parameter type too strict.

This change allows `PyExportTraits` to define `py_typing_param` as an
alternative type-hint to be used for parameter values. `py_typing` then
only servers as the type-hint for return values and constants.
If `py_typing_param` is not defined, then `py_typing` is used for both.

`T!` as strict type:

There's one catch: Callbacks are annotated as `Callable[[P1, ...], R]]`
but the values of the _parameters_ P1 do not flow from Python to C++,
but in the _opposite_ direction: they are being  passed from C++ to the
Python callable function serving as callback. Therefore, they must use
the more strict `py_typing`. Therefore, the exclamation mark can be
used in `py_typing` that a template parameter should never be
substituted using the `py_typing_param` type-hint.