Add MaybeNone as alternative for NoNone
This wrapper in C++ will help marking return values as `T | MaybeNone`.

The typeshed [1] has a helper type `MaybeNone` that's an alias for `Any`
to perform the so-called "Any trick" [2] on a type. This helps
annotating that a type may be `None` without needing to check for `None`
all of the time (assuming that in normal flows it never will be), while
still allowing to _explicitly_ check for it (if the return type was
simply `T`, the type-checker wouldn't allow you to check for `None`).

[1] https://github.com/python/typeshed?tab=readme-ov-file#the-_typeshed-package
[2] https://typing.python.org/en/latest/guides/writing_stubs.html#the-any-trick