Support not-default-constructible types in Python
As we need to default-construct temp variables to get the values out of
Python, we can't use types that don't have a default constructor.
By specializing ArgumentTraits for these types, which we already need to
do to use that as function parameters, we can work around the issue by
using a different storage type.