Use get_parameters in lass_stubgen
Instead of manually filtering the PARAM_DECL children, use
get_parameters() to immedately iterate over the parameter nodes.

This especially helps with the case of std::function return types that
also had its PARM_DECL nodes mixed in, and for which we use the
semantic_parent to filter the real parameters.