Add helpers for Python iterator range views
Returning a PyIteratorRange results in a generic `Iterator[Any]` type in
the type stubs. To correctly capture the value type of the iterator, we
must retain type information in the type returned _to_ Python.

In this commit, we add 5 different helpers for different scenarios:
three for iterator based ranges, and two that use index-based iteration.