Fix overload issues with enum types
If a function is overloaded on both an enum type and another type that's
not convertible to the enum type, and the overload with the enum type is
ranked higher in priority; then the second overload with the other type
will never be considered, because the attempt to match the first
overload will yield in a `ValueError`, effectively terminating the
overload matching:

```
ValueError: x is not a valid E
```

To prevent this, we'll first check if the incoming object is either of
the specific enum type, or an `int` or `str` in case of an
`IntEnumDefinition` or `StrEnumDefinition`. If it's any other type,
we'll immediately raise a `TypeError` so that the overload resolution
can continue.

If it's already of the enum type, then we can immediately retrieve the
value. Otherwise we still need to conver the `int` or `str` to the enum
type to check if it's a correct value. If it's not, then it can still
raise a `ValueError` and terminate the overload resolution, but that's
rather expected: the argument is of the right type, but it has an
invalid value.

So in short, overload resolution works on types, not on values. And you
can't overload an `IntEnum` or `StrEnum` parameter with an `int` or
`str` parameter if the latter is ranked lower in priority: it will never
match.