Type-hint enum parameters as union.
`IntEnum` and `StrEnum` parameters also accept an `int` or `str`
argument that is attempted to construct an enum from, and can raise a
`ValueError` if it's an invalid value, (see previous commit). But `int`
and `str` arguments are of the right type, so let's spell that out when
we type hint these enum types as parameters.