Use TypeInfo for all C++ types in stubdata
In many places, a str was used for a C++ type making it ambigious if it
was just the class name or if it included the template arguments. Since
`TypeInfo` is now frozen and can be used as keys, and since `TypeInfo`
now serializes to just a string if there's only a name (and no arguments
or return value), we can commit on using `TypeInfo` to represent most
C++ types.

There are still a few places where just the type name as key makes sense,
like in the table of export traits and builtin typers.