Fix PyExportTraits for integers
Just like for floating point numbers [1] or complex numbers [2],
PyExportTraits for integers was having the wrong behavior for out of
range values: it would either raise a `TypeError` or even a `SystemError`

We've fixed it to consistently raise an `OverflowError` if the value is
out of range for the destination type (`OverflowError` instead of
`ValueError` to be consistent with Python C API [3,4,5]).

We've changed the explicit `PyLong_Check`. For signed integers, we
removed it, meaning that we start accepting objects with the `__index__`
method that convert them to integers. We're not advertising it with the
`py_typing_param` like `_Float` or `_Complex`, to keep it simple.

For unsigned integers, we still do as `PyLong_AsUnsignedLongLong`
doesn't perform this conversion, so we need to to this explicitly.

For Python <= 3.9, we also still need to perform an explicit
`PyLong_Check` (with `__index__` conversion), because
`PyLong_AsLongLong` would also use the `__int__` method to implicitly
convert objects to integers [6], maning that `float` would also be
accepted, and that we don't want. In Python 3.9, this raises a
deprecation warning, and in Python 3.10 this `__int__` conversion was
dropped [7]:

> DeprecationWarning: an integer is required (got type float).
> Implicit conversion to integers using __int__ is deprecated, and may
> be removed in a future version of Python.

By defining `LASS_USE_OLD_EXPORTTRAITS_INT=1` (in CMake), you can revert
back to the old behavior.

[1] commit dea01c4199933b28fc96b298c4855f2663572ded
[2] commit c6f936971a477243f205c9fe377fe16225055244
[3] https://docs.python.org/3.13/c-api/long.html#c.PyLong_AsLong
[4] https://docs.python.org/3.13/c-api/long.html#c.PyLong_AsUnsignedLongLong
[5] https://docs.python.org/3.13/library/exceptions.html#OverflowError
[6] https://docs.python.org/3.8/c-api/long.html#c.PyLong_AsLongLong
[7] https://docs.python.org/3.10/c-api/long.html#c.PyLong_AsLongLong