Fix (missing) support for generic EnumDefinition
Fix Python parameter type and add missing macros.

Unlike for `StrEnum`, `IntEnum` or `IntFlag`, we do not accept the value
type as argument type for the generic `enum.Enum` case: `StrEnum`
derives from `str` so that their members _are_ `str`, but not in the
`enum.Enum` case: it doesn't derive from the value type and its members
are not _of_ the value type but `enum.Enum`.

Therefore, we fixed the implementation of `doValueObject` to only accept
objects of the enum type. We did that on the base class directly, as
this is the most generic case.

We fixed the stubgen parser and stubdata to store `...` if the value
type cannot be correctly parsed (`Any`)

We fixed the stubgen generator to only add the alternative parameter
types for `StrEnum`, `IntEnum` and `IntFlag`. By using a class-level
dict `_ENUM_ALT_PARAM_TYPES`, this can be expanded when deriving a
custom stub generator.