Add documentation on defining modules
Define a doxygen topic group ModuleDefinition that documents the
ModuleDefinition class and all macros that act on a ModuleDefinition.

Using the following construct, we can make different sections in the
ModuleDefinition topic, grouping different sets of related macros. It
only seems to work with macros, not with functions or classes :-(

  ```
  /** @addtogroup ModuleDefinition
   *  @name Title of the Subsection
   *  some more details
   *  @{
   */
  #define MY_MACRO_1
  #define MY_MACRO_2
  /** @} */
  ```