Fix numCast to always have integers in error
While attempting to add support for `enum.IntFlag`, it was found that
the error message of numCast could cause an encoding issue when trying
to use `PyExportTraits<unsigned char>::get` to get an `int` from Python

The underlying type of the example C++ enum  `Breakfast` was
`unsigned char` and the test was trying to cast 1234 to  it. Because the
boundary=KEEP is being used for the `IntFlag`, this would make it as
value into the `PyExportTraits` for unsigned char and throw a
`BadNumCast` exception with an `unsigned char` value printed as the
character value, not an integer. This would result in an invalid UTF-8
sequence, which the resulting `PyErr_SetString` would eventually fail on,
and we were getting a `ValueError` instead of `OverflowError`.

What we're doing here is to always cast the integers In and Out to a
type that's big enough for both _and_ for an `int` or an `unsigned int`.
Then when we format the error, we'll always cast `in` to that common
type, so that it's never printed as a character.