Fix (missing) support for returning util::Callback* to Python
Until now, we could have a `util::Callback*` parameter and have it wrap a
Python callable, so that we could pass Python callable to C++ functions.

But we couldn't return `util::Callback*` values back to Python as
callables, despite the `py_typing` claiming otherwise: it would simply
return `True`.

First, I tried to drop support for returning `util::Callback*` by
removing the `build` function from its `PyExportTraits`, but that
complicated having `util::Callback*` properties/members that must always
be readable and thus return a callable.

So in the end, we just added support for returning `util::Callback*`,
very similar to how we support it for `std::function`, though maybe a
bit more complicated ...

PS: if you get an error like this, capturing lambdas don't support
`operator!`. Just wrap it in a `std::function` first. See
`testCallbackFromCpp` in `python_embedding.cpp`.

```
built-in operator '!' cannot be applied to an operand of type 'const testCallbackFromCpp::<lambda_1>
```