Use Py_NO_LINK_LIB macro on Windows
Python 3.14 added support for the Py_NO_LINK_LIB [1] macro that you can
use to tell Python to _not_ try to automatically link to the Python
libraries on Windows. Since CMake 3.31.8, CMake will add this macro on
the imported Python libraries [2,3] because it supplies explicit paths
to libraries to link to.

This change adds the same macro to our syspython Conan dependency, to
the same effect.

[1] https://docs.python.org/3.14/extending/windows.html#c.Py_NO_LINK_LIB
[2] https://gitlab.kitware.com/cmake/cmake/-/work_items/26756
[3] https://discourse.cmake.org/t/findpython-option-to-skip-debug-binaries-on-windows/15366