Add unittest on reimporting Python modules
When we reimport a Lass-exported extension module, we expect to get the
same module object as before. This is not a given, even for single-phase
module initialization, but it is true in the way we implemented them.

For simple repeated `import` statement, this is evidently true as Python
will find our module in `sys.modules` and "import" it. This is true for
all Python modules.

But if our module was removed from `sys.modules`, then a subsequent
`import` will re-initialize the module. In case of our single-phase
initialization, that means calling `PyInit_<modulename>` again. This
could create a new module object, but in our case, we return the
internal module object that we still own a reference to.