Disable Garbage Collection on Python heap-types
Stop visiting `ob_type` in `tp_traverse`, and completely disable GC on
the heap-types to fix an access violation in `TestBar.testMapProtocol`
with Python 3.14.

A previous commit 38aed3d04b7663d90b8fd15195704c03ec030842 added a
unittest class `Bar.DerivedInnerClass` to prevent a regression on an
`isFrozen_` check. In Python 3.14 this causes an access violation in the
garbage collector:

  ```
  python314.dll!_PyGCHead_SET_NEXT(PyGC_Head *) Line 168
  python314.dll!gc_list_move(PyGC_Head * node, PyGC_Head * list) Line 302
  python314.dll!mark_stacks(_is * interp, PyGC_Head * visited, int visited_space, bool start) Line 1500
  python314.dll!gc_collect_increment(_ts * tstate, gc_collection_stats * stats) Line 1627
  python314.dll!_PyGC_Collect(_ts * tstate, int generation, _PyGC_Reason reason) Line 2049
  python314.dll!_Py_RunGC(_ts * tstate) Line 2262
  python314.dll!_Py_HandlePending(_ts * tstate) Line 1399
  python314.dll!_PyEval_EvalFrameDefault(_ts * tstate, _PyInterpreterFrame * frame, int throwflag) Line 14
  ```

When we switched to heap-types in 2e3f1e5bdfa0f3da88771628ae7c838abc2f0055
we've followed the advice to enable GC and at least visit `ob_type` [1,2]
because of reference cycles that can exist:
- heap type instances (are supposed to) have a strong reference to
  `ob_type` as the type is no longer static (or immortal) [7]
- heap type have a strong reference to the module they exist, by using
  `PyType_FromModuleAndSpec` to be able to get the module state [6]
- the module can have strong references to heap type instances, and also
  the heap type itself.

But to properly support GC, you must also use `PyObject_GC_New` and
`PyObject_GC_Del` to properly allocate the extra space for the GC meta
data, and use `PyObject_GC_Track` to start tracking it. Or rather [2]:
use `tp_alloc` (which is set to `PyType_GenericAlloc` [3,4]) and
`tp_free` (which is set to `PyObject_GC_Del` [5]).

But we don't: we are simply using C++ `new` (in `tp_new`) and `delete`
(in `tp_dealloc`), and we're not using `tp_alloc` (that would have been
called from `tp_new`) or `tp_free`.

However, as `ClassDefinition` is already holding a strong reference to the
heap-type until the end of the process, we are already "leaking" the
type, and thus preventing this with GC becomes less relevant. In the
same way, `ModuleDefinition` is holding a strong reference to the module
until the end of the process, reusing it on re-import.

This may change when we switch to multi-phase initialization of modules
to properly support subinterpreters. But I expect that we'll not use
heap-types per module (it really complicates determining the right
heap-type for C++ return values), but rather per interpreter. And maybe
also one module per interpreter. Which means we may not need to be able to
get the module state from the type [6], so we don't need to use
`PyType_FromModuleAndSpec()` and break the cycle by not holding a strong
reference to the module in the type, eliminating the need for GC.

Thus for now, we no longer participate in GC with our heap-types, as
we're "leaking" the type anyway. At worst, we're leaking it even more.

An alternative approach could be to overload the C++ `new` and `delete`
operators to use `tp_alloc` and `tp_free`, but I couldn't get it quite
working. I eventually ended up with the same access violation in GC.
Here's a quick sketch:

  ```
  #define PY_HEADER( t_parentClass ) \
    public: \
      /* existing code */ \
      static void* operator new(size_t size) { return mynew(_lassPyClassDef, size); } \
      static void operator delete(void* mem, size_t size) { mydelete(_lassPyClassDef, mem, size); } \
    private:

  void* mynew(ClassDefinition& classDef, size_t size)
  {
    auto t = classDef.type();
    void* mem = t->tp_alloc(t, 0);
    if (!mem) throw std::bad_alloc();
    return mem;
  }

  void mydelete(ClassDefinition& classDef, void* mem, size_t size)
  {
    auto t = classDef.type();
    Py_SET_TYPE(mem, t); // PyObject_GC_Del needs ob_type
    t->tp_free(mem);
  }
  ```

[1] https://docs.python.org/3.14/howto/isolating-extensions.html#heap-types
[2] https://pythondev.readthedocs.io/garbage_collector.html
[3] https://docs.python.org/3.14/c-api/type.html#c.PyType_GenericAlloc
[4] https://docs.python.org/3.14/c-api/typeobj.html#c.PyTypeObject.tp_alloc
[5] https://docs.python.org/3.14/c-api/typeobj.html#c.PyTypeObject.tp_free
[6] https://docs.python.org/3.14/howto/isolating-extensions.html#module-state-access-from-classes
[7] https://bugs.python.org/issue35810