Use PyType_Freeze for Python >= 3.14
This is the proper way to set Py_TPFLAGS_IMMUTABLETYPE and is part of
the stable API.

There is one catch: You can't freeze a type if its parents aren't
immutable yet! [1] It would raise a TypeError (Modifying tp_flags
didn't suffer from this problem, of course).

That's a problem with subclasses that are also inner classes of their
parent. For example, the DerivedInnerClass type in the test suite that
both derives from Bar and is an inner class: Bar.DerivedInnerClass.

When Bar's freezeDefinition is called, it will DerivedInnerClass' type
as inner class, and call DerivedInnerClass' freezeDefinition. But that
one can't call freezeType yet, as Bar's type isn't immutalble yet!

The solution here is to skip freezing the type if the parent isn't
frozen yet, and have the parent freeze all the subclasses that have been
skipped.

[1] https://docs.python.org/3.15/c-api/type.html#c.PyType_Freeze